clean:
* main pom.xml
* lib/core pom.xml
& lib/core packages
