--- databus-mods-lib/databus-mods-core/src/main/scala/lib/databus/DatabusIdentifierType.scala
@@ -1,6 +1,0 @@
-package lib.databus
-
-object DatabusIdentifierType extends Enumeration {
-  type DatabusIdentifierType = Value
-  val PUBLISHER_ID, GROUP_ID, ARTIFACT_ID, VERSION_ID, FILE_ID = Value
-}
