--- databus-mods-lib/databus-mods-core/src/main/scala/lib/model/MetadataExtensionBuilder.scala
@@ -1,5 +1,0 @@
-package lib.model
-
-class MetadataExtensionBuilder {
-
-}
