--- databus-mods-lib/databus-mods-core/src/main/scala/lib/vocab/ModVocabModelBuilder.scala
@@ -1,25 +1,0 @@
-package lib.vocab
-
-import org.apache.jena.rdf.model.{Model, ModelFactory}
-
-class ModVocabModelBuilder {
-
-  private val model = ModelFactory.createDefaultModel()
-
-  def inheritModStatisticsDerivedFrom() : Unit = {
-
-  }
-
-  def inheritModEnrichmentDerivedFrom(): Unit = {
-
-  }
-
-  def inheritModStatistics(): Unit = {
-
-  }
-
-  def build: Model = {
-
-    model
-  }
-}
