--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/AsyncWorker.scala
@@ -1,13 +1,0 @@
-package lib.worker
-
-import org.springframework.context.annotation.{ComponentScan, Configuration}
-
-@Configuration
-@ComponentScan(basePackages = Array(
-  "org.dbpedia.databus_mods.lib.worker.controller",
-  "org.dbpedia.databus_mods.lib.worker.service",
-  "org.dbpedia.databus_mods.lib.worker.execution"
-))
-class AsyncWorker {
-
-}
