--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/controller/IModApiController.scala
@@ -1,5 +1,0 @@
-//package org.dbpedia.databus_mods.lib.worker.controller.impl
-//
-// IModApiController {
-//
-//}
