--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/controller/impl/ModApiController.scala
@@ -1,5 +1,0 @@
-package lib.worker.controller.impl
-
-class ModApiController {
-
-}
