--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/execution/ModProcessor.scala
@@ -1,8 +1,0 @@
-package lib.worker.execution
-
-import org.dbpedia.databus_mods.lib.model.MetadataExtensionBuilder
-
-trait ModProcessor {
-
-  def process(extension: Extension)
-}
