--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/execution/ModRequest.scala
@@ -1,8 +1,0 @@
-package lib.worker.execution
-
-import java.net.URI
-
-case class ModRequest(databusID: String, sourceURI: String) {
-
-  lazy val databusPath: String = new URI(databusID).getPath
-}
