--- databus-mods-lib/databus-mods-core/src/main/scala/lib/worker/execution/ModResultFile.scala
@@ -1,5 +1,0 @@
-package lib.worker.execution
-
-import java.io.File
-
-case class ModResultFile(file: File)
