+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/databus/DatabusIdentifierType.scala
@@ -1,0 +1,6 @@
+package org.dbpedia.databus.mods.core.databus
+
+object DatabusIdentifierType extends Enumeration {
+  type DatabusIdentifierType = Value
+  val PUBLISHER_ID, GROUP_ID, ARTIFACT_ID, VERSION_ID, FILE_ID = Value
+}
