+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/model/MetadataExtension.scala
@@ -1,0 +1,5 @@
+package org.dbpedia.databus.mods.core.model
+
+class MetadataExtension {
+
+}
