+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/vocab/ModVocabModelBuilder.scala
@@ -1,0 +1,25 @@
+package org.dbpedia.databus.mods.core.vocab
+
+import org.apache.jena.rdf.model.{Model, ModelFactory}
+
+class ModVocabModelBuilder {
+
+  private val model = ModelFactory.createDefaultModel()
+
+  def inheritModStatisticsDerivedFrom() : Unit = {
+
+  }
+
+  def inheritModEnrichmentDerivedFrom(): Unit = {
+
+  }
+
+  def inheritModStatistics(): Unit = {
+
+  }
+
+  def build: Model = {
+
+    model
+  }
+}
