+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/worker/AsyncWorker.scala
@@ -1,0 +1,13 @@
+package org.dbpedia.databus.mods.core.worker
+
+import org.springframework.context.annotation.{ComponentScan, Configuration}
+
+@Configuration
+@ComponentScan(basePackages = Array(
+  "org.dbpedia.databus_mods.lib.worker.controller",
+  "org.dbpedia.databus_mods.lib.worker.service",
+  "org.dbpedia.databus_mods.lib.worker.execution"
+))
+class AsyncWorker {
+
+}
