+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/worker/execution/ModProcessor.scala
@@ -1,0 +1,6 @@
+package org.dbpedia.databus.mods.core.worker.execution
+
+trait ModProcessor {
+
+  def process(extension: Extension)
+}
