+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/worker/execution/ModRequest.scala
@@ -1,0 +1,8 @@
+package org.dbpedia.databus.mods.core.worker.execution
+
+import java.net.URI
+
+case class ModRequest(databusID: String, sourceURI: String) {
+
+  lazy val databusPath: String = new URI(databusID).getPath
+}
