+++ databus-mods-lib/databus-mods-core/src/main/scala/org/dbpedia/databus/mods/core/worker/execution/ModResultFile.scala
@@ -1,0 +1,5 @@
+package org.dbpedia.databus.mods.core.worker.execution
+
+import java.io.File
+
+case class ModResultFile(file: File)
