+++ databus-mods-lib/databus-mods-spring-boot-starter/${project.build.directory}/classes/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.dbpedia.databus.mods.worker.springboot.config.ModWorkerApiAutoConfig
