+++ databus-mods-lib/pom.xml
@@ -4,91 +4,82 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>databus-mods-lib</artifactId>
-    <packaging>pom</packaging>
-    <name>Databus-Mods Core-Library</name>
-    <modules>
-        <module>databus-mods-core</module>
-        <module>databus-mods-spring-boot-starter</module>
-    </modules>
-
     <parent>
         <groupId>org.dbpedia.databus-mods</groupId>
         <artifactId>databus-mods-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <repositories>
-        <repository>
-            <id>jitpack.io</id>
-            <url>https://jitpack.io</url>
-        </repository>
-    </repositories>
+    <artifactId>databus-mods-lib</artifactId>
+    <packaging>pom</packaging>
+    <name>Databus-Mods Core-Library</name>
 
-    <dependencies>
+    <modules>
+        <module>databus-mods-core</module>
+<!--        <module>databus-mods-spring-boot-starter</module>-->
+    </modules>
+
+<!--    <dependencies>-->
+
+<!--        &lt;!&ndash; JENA &ndash;&gt;-->
 <!--        <dependency>-->
-<!--            <groupId>com.github.Vehnem</groupId>-->
-<!--            <artifactId>rdfdetector</artifactId>-->
-<!--            <version>master-SNAPSHOT</version>-->
+<!--            <groupId>org.apache.jena</groupId>-->
+<!--            <artifactId>jena-shacl</artifactId>-->
+<!--            <version>3.17.0</version>-->
+<!--            <exclusions>-->
+<!--                <exclusion>-->
+<!--                    <groupId>com.github.jsonld-java</groupId>-->
+<!--                    <artifactId>jsonld-java</artifactId>-->
+<!--                </exclusion>-->
+<!--            </exclusions>-->
 <!--        </dependency>-->
-        <!-- https://mvnrepository.com/artifact/org.apache.jena/jena-shacl -->
-        <dependency>
-            <groupId>org.apache.jena</groupId>
-            <artifactId>jena-shacl</artifactId>
-            <version>3.17.0</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.github.jsonld-java</groupId>
-                    <artifactId>jsonld-java</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.jena</groupId>
-            <artifactId>jena-arq</artifactId>
-            <version>3.17.0</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.github.jsonld-java</groupId>
-                    <artifactId>jsonld-java</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.jena</groupId>
-            <artifactId>jena-core</artifactId>
-            <version>3.17.0</version>
-            <exclusions>
-            <exclusion>
-                <groupId>com.github.jsonld-java</groupId>
-                <artifactId>jsonld-java</artifactId>
-            </exclusion>
-            </exclusions>
-        </dependency>
-<!--        &lt;!&ndash; https://mvnrepository.com/artifact/javax.persistence/javax.persistence-api &ndash;&gt;-->
 <!--        <dependency>-->
-<!--            <groupId>javax.persistence</groupId>-->
-<!--            <artifactId>javax.persistence-api</artifactId>-->
-<!--            <version>2.2</version>-->
+<!--            <groupId>org.apache.jena</groupId>-->
+<!--            <artifactId>jena-arq</artifactId>-->
+<!--            <version>3.17.0</version>-->
+<!--            <exclusions>-->
+<!--                <exclusion>-->
+<!--                    <groupId>com.github.jsonld-java</groupId>-->
+<!--                    <artifactId>jsonld-java</artifactId>-->
+<!--                </exclusion>-->
+<!--            </exclusions>-->
+<!--        </dependency>-->
+<!--        <dependency>-->
+<!--            <groupId>org.apache.jena</groupId>-->
+<!--            <artifactId>jena-core</artifactId>-->
+<!--            <version>3.17.0</version>-->
+<!--            <exclusions>-->
+<!--            <exclusion>-->
+<!--                <groupId>com.github.jsonld-java</groupId>-->
+<!--                <artifactId>jsonld-java</artifactId>-->
+<!--            </exclusion>-->
+<!--            </exclusions>-->
 <!--        </dependency>-->
 
+<!--        &lt;!&ndash; CLI &ndash;&gt;-->
 <!--        <dependency>-->
-<!--            <groupId>org.springframework.boot</groupId>-->
-<!--            <artifactId>spring-boot-starter-data-jpa</artifactId>-->
-<!--            <version>2.3.0.RELEASE</version>-->
+<!--            <groupId>info.picocli</groupId>-->
+<!--            <artifactId>picocli-spring-boot-starter</artifactId>-->
+<!--            <version>4.6.1</version>-->
 <!--        </dependency>-->
-
-<!--        <dependency>-->
-<!--            <groupId>org.apache.derby</groupId>-->
-<!--            <artifactId>derby</artifactId>-->
-<!--            <version>10.14.2.0</version>-->
-<!--        </dependency>-->
-
-        <dependency>
-            <groupId>info.picocli</groupId>
-            <artifactId>picocli-spring-boot-starter</artifactId>
-            <version>4.6.1</version>
-        </dependency>
-    </dependencies>
-
+<!--    </dependencies>-->
 </project>
+
+        <!--        &lt;!&ndash; https://mvnrepository.com/artifact/javax.persistence/javax.persistence-api &ndash;&gt;-->
+        <!--        <dependency>-->
+        <!--            <groupId>javax.persistence</groupId>-->
+        <!--            <artifactId>javax.persistence-api</artifactId>-->
+        <!--            <version>2.2</version>-->
+        <!--        </dependency>-->
+
+        <!--        <dependency>-->
+        <!--            <groupId>org.springframework.boot</groupId>-->
+        <!--            <artifactId>spring-boot-starter-data-jpa</artifactId>-->
+        <!--            <version>2.3.0.RELEASE</version>-->
+        <!--        </dependency>-->
+
+        <!--        <dependency>-->
+        <!--            <groupId>org.apache.derby</groupId>-->
+        <!--            <artifactId>derby</artifactId>-->
+        <!--            <version>10.14.2.0</version>-->
+        <!--        </dependency>-->
