+++ pom.xml
@@ -2,11 +2,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>2.3.7.RELEASE</version>
+        <relativePath/> <!-- lookup parent from repository -->
+    </parent>
+
     <groupId>org.dbpedia.databus-mods</groupId>
     <artifactId>databus-mods-parent</artifactId>
+    <version>1.0-SNAPSHOT</version>
     <name>Databus-Mods Parent</name>
     <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
 
     <properties>
         <encoding>UTF-8</encoding>
@@ -25,18 +32,12 @@
         <maven.compiler.source>1.8</maven.compiler.source>
     </properties>
 
-    <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.7.RELEASE</version>
-        <relativePath/> <!-- lookup parent from repository -->
-    </parent>
-
     <modules>
         <module>databus-mods-lib</module>
-        <module>databus-mods-server</module>
+<!--        <module>databus-mods-server</module>-->
     </modules>
 
+    <dependencyManagement>
     <dependencies>
 
         <!--
@@ -58,6 +59,25 @@
         END This is from mods pom.xml
         -->
 
+        <dependency>
+            <groupId>org.apache.jena</groupId>
+            <artifactId>jena-core</artifactId>
+            <version>3.17.0</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.jena</groupId>
+            <artifactId>jena-arq</artifactId>
+            <version>3.17.0</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.jena</groupId>
+            <artifactId>jena-shacl</artifactId>
+            <version>3.17.0</version>
+        </dependency>
+
+
         <!-- https://mvnrepository.com/artifact/org.scala-lang.modules/scala-java8-compat -->
         <dependency>
             <groupId>org.scala-lang.modules</groupId>
@@ -73,6 +93,7 @@
         </dependency>
 
 
+        <!-- Spring -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
@@ -90,7 +111,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-thymeleaf</artifactId>
@@ -102,8 +122,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-
-
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
@@ -134,9 +152,7 @@
             <version>3.0.0</version>
         </dependency>
 
-        <!-- config -->
-
-        <!-- databus client -->
+        <!-- Databus Client -->
         <dependency>
             <groupId>com.github.dbpedia</groupId>
             <artifactId>databus-client</artifactId>
@@ -174,6 +190,7 @@
         </dependency>
 
     </dependencies>
+    </dependencyManagement>
 
     <repositories>
         <!--        <repository>-->
@@ -194,16 +211,16 @@
             <id>sonatype-releases</id>
             <url>https://oss.sonatype.org/service/local/repositories/releases/content/</url>
         </repository>
-        <repository>
-            <id>archiva.internal</id>
-            <name>Internal Release Repository</name>
-            <url>http://databus.dbpedia.org:8081/repository/internal</url>
-        </repository>
-        <repository>
-            <id>archiva.snapshots</id>
-            <name>Internal Snapshot Repository</name>
-            <url>http://databus.dbpedia.org:8081/repository/snapshots</url>
-        </repository>
+<!--        <repository>-->
+<!--            <id>archiva.internal</id>-->
+<!--            <name>Internal Release Repository</name>-->
+<!--            <url>https://databus.dbpedia.org:8081/repository/internal</url>-->
+<!--        </repository>-->
+<!--        <repository>-->
+<!--            <id>archiva.snapshots</id>-->
+<!--            <name>Internal Snapshot Repository</name>-->
+<!--            <url>https://databus.dbpedia.org:8081/repository/snapshots</url>-->
+<!--        </repository>-->
     </repositories>
 
     <distributionManagement>
@@ -212,16 +229,16 @@
             <name>Temporary Staging Repository</name>
             <url>file://${project.build.directory}/mvn-repo</url>
         </repository-->
-        <repository>
-            <id>archiva.internal</id>
-            <name>Internal Release Repository</name>
-            <url>http://databus.dbpedia.org:8081/repository/internal/</url>
-        </repository>
-        <snapshotRepository>
-            <id>archiva.snapshots</id>
-            <name>Internal Snapshot Repository</name>
-            <url>http://databus.dbpedia.org:8081/repository/snapshots/</url>
-        </snapshotRepository>
+<!--        <repository>-->
+<!--            <id>archiva.internal</id>-->
+<!--            <name>Internal Release Repository</name>-->
+<!--            <url>https://databus.dbpedia.org:8081/repository/internal/</url>-->
+<!--        </repository>-->
+<!--        <snapshotRepository>-->
+<!--            <id>archiva.snapshots</id>-->
+<!--            <name>Internal Snapshot Repository</name>-->
+<!--            <url>https://databus.dbpedia.org:8081/repository/snapshots/</url>-->
+<!--        </snapshotRepository>-->
     </distributionManagement>
 
     <build>
@@ -240,7 +257,7 @@
             <plugin>
                 <groupId>net.alchim31.maven</groupId>
                 <artifactId>scala-maven-plugin</artifactId>
-                <version>3.2.0</version>
+                <version>4.5.4</version>
                 <executions>
                     <execution>
                         <goals>
@@ -253,6 +270,7 @@
                                 <arg>-dependencyfile</arg>
                                 <arg>${project.build.directory}/.scala_dependencies</arg>
                                 <arg>-deprecation</arg>
+                                <arg>-nobootcp</arg>
                                 <arg>-feature</arg>
                                 <arg>-target:jvm-1.8</arg>
                             </args>
@@ -266,13 +284,13 @@
         </plugins>
     </build>
 
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.dbpedia.databus_mods</groupId>
-                <artifactId>databus-mods-lib</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
+<!--    <dependencyManagement>-->
+<!--        <dependencies>-->
+<!--            <dependency>-->
+<!--                <groupId>org.dbpedia.databus_mods</groupId>-->
+<!--                <artifactId>databus-mods-lib</artifactId>-->
+<!--                <version>${project.version}</version>-->
+<!--            </dependency>-->
+<!--        </dependencies>-->
+<!--    </dependencyManagement>-->
 </project>
